AirconBot works
The aircon thingy works but i believe i am wasting time by doing this and not working on the library bot. This will the be last commit on this for a while.The project isnt perfect but i will have to update it when i start wasting time once class starts.